feat(workflows): update Check-releases.yml and build-telegram.yml
Updated the Check-releases.yml file to include an additional repository (Xposed-Modules-Repo/com.simo.tgramhooks) in the list of repositories to check for new releases.

In the build-telegram.yml file, uncommented the line to enable a module (TG.H.*.apk) by adding the closing quotation mark that was missing.

These changes ensure that the new repository is included in the release check process and enable the specified module during the build process.